move views to related files
