+++ api/views/__init__.py
@@ -1,0 +1,2 @@
+from .social_view import MemberViewSet, TeamViewSet
+from .ticket_view import TicketViewSet, TicketCommentViewSet, TicketStatusViewSet
