+++ api/views/social_view.py
@@ -4,23 +4,17 @@ from rest_framework.response import Response
 from django.utils import timezone
 from rest_framework import exceptions
 
-from django.contrib.auth import get_user_model, get_user
+from django.contrib.auth import get_user_model
 
-from .models import (
+from ..models import (
     Member,
     Team,
-    Ticket,
-    TicketComment,
-    TicketStatus
 )
-from .serializers import (
-    TicketSerializer,
+from ..serializers import (
     MemberSerializer,
     TeamSerializer,
-    TicketCommentSerializer,
     UserSerializer
 )
-from .permissions import IsOwnerOrReadOnly, IsAdminOrReadOnly
 
 
 class MemberViewSet(viewsets.ModelViewSet):
@@ -142,54 +136,3 @@ class TeamViewSet(viewsets.ModelViewSet):
             cosine similarity
         """
         pass
-
-
-class TicketViewSet(viewsets.ModelViewSet):
-    """
-    This viewset automatically provides `list` and `detail` actions.
-    """
-
-    permission_classes = [
-        permissions.IsAuthenticatedOrReadOnly,
-        # IsOwnerOrReadOnly | IsAdminOrReadOnly, ignoring these until we get owner field squared away
-    ]
-
-    queryset = Ticket.objects.all()
-    serializer_class = TicketSerializer
-
-    @action(detail=False, methods=["get"])
-    def list_team(self, request, pk=None):
-        queryset = Ticket.objects.filter(team__id=pk)
-        serializer = self.serializer_class(queryset, many=True)
-        return Response(serializer.data)
-
-    def perform_create(self, serializer):
-        serializer.save(owner=self.request.user)
-
-
-class TicketCommentViewSet(viewsets.ModelViewSet):
-    """
-    Basic crud should be pre-generated, so we only need to do the more complicated calls
-    """
-
-    permission_classes = [
-        permissions.IsAuthenticatedOrReadOnly,
-        IsOwnerOrReadOnly | IsAdminOrReadOnly,
-    ]
-
-    queryset = TicketComment.objects.all()
-    serializer_class = TicketCommentSerializer
-
-    @action(detail=False)
-    def recent_comments(self, request, team_id=None):
-        """ This call returns the first page of comments associated with the given team_id """
-        pass
-
-
-class TicketStatusViewSet(viewsets.ModelViewSet):
-    permission_classes = [
-        permissions.IsAuthenticatedOrReadOnly,
-        IsOwnerOrReadOnly | IsAdminOrReadOnly
-    ]
-
-    queryset = TicketStatus.objects.all()
